* création des méthodes helper pour le mv ou cp de répertoires (commons io). Tests junit.

+++ ornidroid/src/main/java/fr/ornidroid/helper/FileHelper.java
@@ -1,13 +1,28 @@
 package fr.ornidroid.helper;
 
 import java.io.File;
+import java.io.FileInputStream;
 import java.io.FileNotFoundException;
+import java.io.FileOutputStream;
 import java.io.IOException;
+import java.nio.channels.FileChannel;
+import java.util.ArrayList;
+import java.util.List;
 
 /**
- * The Class FileUtils. code from commons-io 1.4
+ * The Class FileUtils. code from commons-io 2.4
  */
 public class FileHelper {
+
+	/** The Constant ONE_KB. */
+	public static final long ONE_KB = 1024;
+
+	/** The Constant ONE_MB. */
+	public static final long ONE_MB = ONE_KB * ONE_KB;
+
+	/** The Constant FILE_COPY_BUFFER_SIZE. */
+	private static final long FILE_COPY_BUFFER_SIZE = ONE_MB * 30;
+
 	/**
 	 * Creates the empty file if it does not exist.
 	 * 
@@ -40,17 +55,19 @@ public class FileHelper {
 	 *             in case deletion is unsuccessful
 	 */
 	public static void forceDelete(final File file) throws IOException {
-		if (file.isDirectory()) {
-			deleteDirectory(file);
-		} else {
-			final boolean filePresent = file.exists();
-			if (!file.delete()) {
-				if (!filePresent) {
-					throw new FileNotFoundException("File does not exist: "
-							+ file);
+		if (file.exists()) {
+			if (file.isDirectory()) {
+				deleteDirectory(file);
+			} else {
+				final boolean filePresent = file.exists();
+				if (!file.delete()) {
+					if (!filePresent) {
+						throw new FileNotFoundException("File does not exist: "
+								+ file);
+					}
+					final String message = "Unable to delete file: " + file;
+					throw new IOException(message);
 				}
-				final String message = "Unable to delete file: " + file;
-				throw new IOException(message);
 			}
 		}
 	}
@@ -79,6 +96,123 @@ public class FileHelper {
 				throw new IOException(message);
 			}
 		}
+	}
+
+	/**
+	 * Moves a directory.
+	 * <p>
+	 * When the destination directory is on another file system, do a
+	 * "copy and delete".
+	 * 
+	 * @param srcDir
+	 *            the directory to be moved
+	 * @param destDir
+	 *            the destination directory
+	 * @throws NullPointerException
+	 *             if source or destination is {@code null}
+	 * @throws FileExistsException
+	 *             if the destination directory exists
+	 * @throws IOException
+	 *             if source or destination is invalid
+	 * @throws IOException
+	 *             if an IO error occurs moving the file
+	 * @since 1.4
+	 */
+	public static void moveDirectory(final File srcDir, final File destDir)
+			throws IOException {
+		if (srcDir == null) {
+			throw new NullPointerException("Source must not be null");
+		}
+		if (destDir == null) {
+			throw new NullPointerException("Destination must not be null");
+		}
+		if (!srcDir.exists()) {
+			throw new FileNotFoundException("Source '" + srcDir
+					+ "' does not exist");
+		}
+		if (!srcDir.isDirectory()) {
+			throw new IOException("Source '" + srcDir + "' is not a directory");
+		}
+		if (destDir.exists()) {
+			throw new IOException("Destination '" + destDir
+					+ "' already exists");
+		}
+		final boolean rename = srcDir.renameTo(destDir);
+		if (!rename) {
+			if (destDir.getCanonicalPath()
+					.startsWith(srcDir.getCanonicalPath())) {
+				throw new IOException("Cannot move directory: " + srcDir
+						+ " to a subdirectory of itself: " + destDir);
+			}
+			copyDirectory(srcDir, destDir);
+			deleteDirectory(srcDir);
+			if (srcDir.exists()) {
+				throw new IOException("Failed to delete original directory '"
+						+ srcDir + "' after copy to '" + destDir + "'");
+			}
+		}
+	}
+
+	/**
+	 * Copies a whole directory to a new location.
+	 * <p>
+	 * This method copies the specified directory and all its child directories
+	 * and files to the specified destination. The destination is the new
+	 * location and name of the directory.
+	 * <p>
+	 * The destination directory is created if it does not exist. If the
+	 * destination directory did exist, then this method merges the source with
+	 * the destination, with the source taking precedence.
+	 * <p>
+	 * <strong>Note:</strong> This method tries to preserve the files' last
+	 * modified date/times using {@link File#setLastModified(long)}, however it
+	 * is not guaranteed that those operations will succeed. If the modification
+	 * operation fails, no indication is provided.
+	 * 
+	 * @param srcDir
+	 *            an existing directory to copy, must not be {@code null}
+	 * @param destDir
+	 *            the new directory, must not be {@code null}
+	 * @throws IOException
+	 *             if an IO error occurs during copying
+	 * @since 1.1
+	 */
+	protected static void copyDirectory(final File srcDir, final File destDir)
+			throws IOException {
+		if (srcDir == null) {
+			throw new NullPointerException("Source must not be null");
+		}
+		if (destDir == null) {
+			throw new NullPointerException("Destination must not be null");
+		}
+		if (srcDir.exists() == false) {
+			throw new FileNotFoundException("Source '" + srcDir
+					+ "' does not exist");
+		}
+		if (srcDir.isDirectory() == false) {
+			throw new IOException("Source '" + srcDir
+					+ "' exists but is not a directory");
+		}
+		if (srcDir.getCanonicalPath().equals(destDir.getCanonicalPath())) {
+			throw new IOException("Source '" + srcDir + "' and destination '"
+					+ destDir + "' are the same");
+		}
+
+		// Cater for destination being directory within the source directory
+		// (see IO-141)
+		List<String> exclusionList = null;
+		if (destDir.getCanonicalPath().startsWith(srcDir.getCanonicalPath())) {
+			final File[] srcFiles = srcDir.listFiles();
+
+			if ((srcFiles != null) && (srcFiles.length > 0)) {
+				exclusionList = new ArrayList<String>(srcFiles.length);
+				for (final File srcFile : srcFiles) {
+					final File copiedFile = new File(destDir, srcFile.getName());
+					exclusionList.add(copiedFile.getCanonicalPath());
+				}
+			}
+		}
+		doCopyDirectory(srcDir, destDir, exclusionList);
 	}
 
 	/**
@@ -139,6 +273,105 @@ public class FileHelper {
 			final String message = "Unable to delete directory " + directory
 					+ ".";
 			throw new IOException(message);
+		}
+	}
+
+	/**
+	 * Internal copy directory method.
+	 * 
+	 * @param srcDir
+	 *            the validated source directory, must not be {@code null}
+	 * @param destDir
+	 *            the validated destination directory, must not be {@code null}
+	 * @param exclusionList
+	 *            List of files and directories to exclude from the copy, may be
+	 *            null
+	 * @throws IOException
+	 *             if an error occurs
+	 * @since 1.1
+	 */
+	private static void doCopyDirectory(final File srcDir, final File destDir,
+
+	final List<String> exclusionList) throws IOException {
+		// recurse
+		final File[] srcFiles = srcDir.listFiles();
+		if (srcFiles == null) { // null if abstract pathname does not denote a
+								// directory, or if an I/O error occurs
+			throw new IOException("Failed to list contents of " + srcDir);
+		}
+		if (destDir.exists()) {
+			if (destDir.isDirectory() == false) {
+				throw new IOException("Destination '" + destDir
+						+ "' exists but is not a directory");
+			}
+		} else {
+			if (!destDir.mkdirs() && !destDir.isDirectory()) {
+				throw new IOException("Destination '" + destDir
+						+ "' directory cannot be created");
+			}
+		}
+		if (destDir.canWrite() == false) {
+			throw new IOException("Destination '" + destDir
+					+ "' cannot be written to");
+		}
+		for (final File srcFile : srcFiles) {
+			final File dstFile = new File(destDir, srcFile.getName());
+			if ((exclusionList == null)
+					|| !exclusionList.contains(srcFile.getCanonicalPath())) {
+				if (srcFile.isDirectory()) {
+					doCopyDirectory(srcFile, dstFile, exclusionList);
+				} else {
+					doCopyFile(srcFile, dstFile);
+				}
+			}
+		}
+
+	}
+
+	/**
+	 * Internal copy file method.
+	 * 
+	 * @param srcFile
+	 *            the validated source file, must not be {@code null}
+	 * @param destFile
+	 *            the validated destination file, must not be {@code null}
+	 * @throws IOException
+	 *             if an error occurs
+	 */
+	private static void doCopyFile(final File srcFile, final File destFile)
+			throws IOException {
+		if (destFile.exists() && destFile.isDirectory()) {
+			throw new IOException("Destination '" + destFile
+					+ "' exists but is a directory");
+		}
+
+		FileInputStream fis = null;
+		FileOutputStream fos = null;
+		FileChannel input = null;
+		FileChannel output = null;
+		try {
+			fis = new FileInputStream(srcFile);
+			fos = new FileOutputStream(destFile);
+			input = fis.getChannel();
+			output = fos.getChannel();
+			final long size = input.size();
+			long pos = 0;
+			long count = 0;
+			while (pos < size) {
+				count = (size - pos) > FILE_COPY_BUFFER_SIZE ? FILE_COPY_BUFFER_SIZE
+						: size - pos;
+				pos += output.transferFrom(input, pos, count);
+			}
+		} finally {
+			IOHelper.closeQuietly(output);
+			IOHelper.closeQuietly(fos);
+			IOHelper.closeQuietly(input);
+			IOHelper.closeQuietly(fis);
+		}
+
+		if (srcFile.length() != destFile.length()) {
+			throw new IOException("Failed to copy full contents from '"
+					+ srcFile + "' to '" + destFile + "'");
 		}
 	}
